refactor: Update task model structure and user model
- Rename task_model.dart to task.dart for consistency
- Add proper ID fields to Task and SubTask models
- Improve model serialization and null safety
- Update User model with enhanced fields
