+++ lib/models/user.dart
@@ -1,0 +1,15 @@
+class User {
+  final int id;
+  final String username;
+  final String email;
+
+  User({required this.id, required this.username, required this.email});
+
+  factory User.fromJson(Map<String, dynamic> json) {
+    return User(
+      id: json['id'],
+      username: json['username'],
+      email: json['email'],
+    );
+  }
+}
